feat(styled-system): enhance folder structure for better readability
