+++ packages/styled-system/src/config/outline.js
@@ -1,7 +1,5 @@
 import system from '../core/system';
-import {
-  positiveOrNegative as positiveOrNegativeTransform,
-} from '../utils/transforms';
+import positiveOrNegativeTransform from '../transforms/positiveOrNegative';
 
 const group = 'outline';
 const config = {
