+++ packages/styled-system/src/config/shape.js
@@ -1,5 +1,5 @@
 import system from '../core/system';
-import { positiveOrNegative as positiveOrNegativeTransform } from '../utils/transforms';
+import positiveOrNegativeTransform from '../transforms/positiveOrNegative';
 
 /**
  * https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_Shapes
