--- packages/styled-system/src/utils/ensure-array.js
@@ -1,9 +1,0 @@
-const ensureArray = (value, defaultValue = []) => {
-  if (value === undefined || value === null) {
-    return [].concat(defaultValue);
-  }
-
-  return Array.isArray(value) ? value : [].concat(value);
-};
-
-export default ensureArray;
